Move components and resources to module
